Move account derivations and add RewardManagerState struct
